Changes to npc-server: - Added default Control-NPC to npcs folder. - Added /savenpcs command to save npcs to the disk. - Log script exceptions to file, and send to NC.

+++ server/src/script/V8LevelImpl.cpp
@@ -57,10 +57,10 @@ void Level_Function_FindAreaNpcs(const v8::FunctionCallbackInfo<v8::Value>& args
 	v8::Local<v8::Context> context = isolate->GetCurrentContext();
 
 	// Argument parsing
-	int startX = 16 * args[0]->NumberValue(context).ToChecked();
-	int startY = 16 * args[1]->NumberValue(context).ToChecked();
-	int endX = 16 * args[2]->IntegerValue(context).ToChecked();
-	int endY = 16 * args[3]->IntegerValue(context).ToChecked();
+	int startX = (int)(16 * args[0]->NumberValue(context).ToChecked());
+	int startY = (int)(16 * args[1]->NumberValue(context).ToChecked());
+	int endX = 16 * args[2]->Int32Value(context).ToChecked();
+	int endY = 16 * args[3]->Int32Value(context).ToChecked();
 
 	// Unwrap Object
 	TLevel *levelObject = UnwrapObject<TLevel>(args.This());
